split todo file
