+++ front-end/virtual_pet/lib/flame/components/todo/todo_item.dart
@@ -1,0 +1,9 @@
+// Todo item class
+class Todo {
+  String name;
+  bool checked;
+
+  Todo({
+    required this.name, 
+    required this.checked});
+}
