+++ front-end/virtual_pet/lib/flame/components/todo/todo_tile.dart
@@ -1,14 +1,5 @@
 import 'package:flutter/material.dart';
-
-// Todo item class
-class Todo {
-  String name;
-  bool checked;
-
-  Todo({
-    required this.name, 
-    required this.checked});
-}
+import 'package:virtual_pet/flame/components/todo/todo_item.dart';
 
 // Todo item tile widget
 class TodoItemTile extends StatefulWidget {
@@ -102,96 +93,6 @@ class _TodoItemTileState extends State<TodoItemTile> {
     return const TextStyle(
       color: Colors.black54,
       decoration: TextDecoration.lineThrough,
-    );
-  }
-}
-
-// List view
-class TodoList extends StatefulWidget {
-  const TodoList({super.key});
-
-  @override
-  State<TodoList> createState() => _TodoListState();
-}
-
-class _TodoListState extends State<TodoList> {
-  final TextEditingController _textFieldController = TextEditingController();
-  final List<Todo> _todos = <Todo>[];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Todo list'),
-      ),
-      body: ListView(
-        padding: const EdgeInsets.symmetric(vertical: 8.0),
-        // populate _todos and display each item as list tile
-        children: _todos.map((Todo todo) {
-          return TodoItemTile(
-            todo: todo,
-            onChecked: _handleChecked,
-            onChanged: _handeChange,
-          );
-        }).toList(),
-      ),
-      floatingActionButton: FloatingActionButton(
-          onPressed: () => _displayDialog(),
-          tooltip: 'Add Item',
-          child: const Icon(Icons.add)),
-    );
-  }
-
-  void _handleChecked(Todo todo) {
-    setState(() {
-      todo.checked = !todo.checked;
-    });
-  }
-
-  void _handeChange(Todo updatedTodo) {
-  }
-  
-  void _addTodoItem(String name) {
-    setState(() {
-      _todos.add(Todo(name: name, checked: false));
-    });
-    _textFieldController.clear();
-  }
-
-  Future<void> _displayDialog() async {
-    return showDialog<void>(
-      context: context,
-      barrierDismissible: false, // user must tap button!
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: const Text('Add a new todo item'),
-          content: TextField(
-            controller: _textFieldController,
-            decoration: const InputDecoration(hintText: 'Type your new todo'),
-          ),
-          actions: <Widget>[
-            TextButton(
-              child: const Text('Add'),
-              onPressed: () {
-                Navigator.of(context).pop();
-                _addTodoItem(_textFieldController.text);
-              },
-            ),
-          ],
-        );
-      },
-    );
-  }
-}
-
-class TodoApp extends StatelessWidget {
-  const TodoApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return const MaterialApp(
-      title: 'Todo list',
-      home: TodoList(),
     );
   }
 }
