+++ front-end/virtual_pet/lib/pages/todo.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+import 'package:virtual_pet/flame/components/todo/todo_list.dart';
+
+class TodoApp extends StatelessWidget {
+  const TodoApp({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const MaterialApp(
+      title: 'Todo list',
+      home: TodoList(),
+    );
+  }
+}
